# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/token/{tokenSymbol}/history": {
            "get": {
                "description": "Retrieve 1-year price history for a specific LST token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tokens"
                ],
                "summary": "Get price history for a token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token symbol (e.g., wstETH, rETH)",
                        "name": "tokenSymbol",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "price_history: array of price points, count: number of data points",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error: invalid token symbol",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error: failed to fetch price data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/token/{tokenSymbol}/valuation": {
            "get": {
                "description": "Retrieve APR, stability, TVL, and valuation remarks for a specific LST token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tokens"
                ],
                "summary": "Get valuation metrics for a token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token symbol (e.g., wstETH, rETH)",
                        "name": "tokenSymbol",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "valuation metrics for the token",
                        "schema": {
                            "$ref": "#/definitions/services.ValuationData"
                        }
                    },
                    "400": {
                        "description": "error: invalid token symbol",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error: failed to calculate valuation",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/tokens": {
            "get": {
                "description": "Retrieve a list of all active Liquid Staking Tokens being tracked",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tokens"
                ],
                "summary": "Get all tracked LST tokens",
                "responses": {
                    "200": {
                        "description": "tokens: array of token objects, count: number of tokens",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "error: error message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/valuations": {
            "get": {
                "description": "Retrieve APR, stability, TVL, and valuation remarks for all tracked LST tokens (sortable table data)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tokens"
                ],
                "summary": "Get valuation metrics for all tokens",
                "responses": {
                    "200": {
                        "description": "valuations: array of valuation objects, count: number of valuations",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "error: failed to fetch valuations",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "services.ValuationData": {
            "type": "object",
            "properties": {
                "apr": {
                    "type": "number"
                },
                "last_updated": {
                    "type": "string"
                },
                "remarks": {
                    "type": "string"
                },
                "stability": {
                    "type": "number"
                },
                "token_symbol": {
                    "type": "string"
                },
                "tvl": {
                    "type": "number"
                }
            }
        }
    }
}